+++ app/models/discount.go
@@ -5,6 +5,6 @@ import "time"
 type Discount struct {
 	ID         string    `gorm:"column:id;primaryKey"`
 	Title      string    `gorm:"column:title;not null"`
-	Discount   uint      `gorm:"column:discount;not null"`
+	Discount   float64   `gorm:"column:discount;not null"`
 	ExpireDate time.Time `gorm:"column:expire_date;not null"`
 }
